Structured the folder Structure
